requirements.txt
streamlit>=1.36.0
faiss-cpu>=1.8.0
numpy>=1.26
pandas>=2.2
openai>=1.40
anthropic>=0.34
pyairtable>=2.3
requests>=2.32